+++ CHANGELOG.rst
@@ -1,8 +1,9 @@
 Jinja Filters Changelog
-===============
+=======================
 
 This Changelog is for ``minchin.pelican.jinja_filters``.
 
+- :support:`7` point to new plugin location at ``pelican-jinja-filters`` on PyPI
 - :release:`1.0.4 <2017-04-17>`
 - :bug:`-` upgrade release machinery
 - :bug:`-` add Pelican trove classifier